backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.20
edge-tts==6.1.12
moviepy==2.1.1
pillow==10.4.0
aiofiles==24.1.0
httpx==0.28.1
google-generativeai==0.8.3
pydantic==2.10.4
pydantic-settings==2.7.0
alembic==1.14.0
